Backend/requirements.txt
pytelegrambotapi~=4.21.0
requests~=2.32.3
uvicorn
fastapi
sqlalchemy
asyncpg
python-jose[cryptography]
python-multipart
aio_pika